requirements.txt
# FinSight AI - 智能金融分析系统
# 基于 LangChain + LangGraph 的对话式 Agent 实现

# Core LangChain packages
langchain==1.1.0
langchain-core==1.1.0
langchain-openai==1.1.0
langchain-anthropic==1.0.0
langchain-community==0.4
langchain-text-splitters==1.0.0
langchain-classic==1.0.0

# LangGraph (现代 Agent 框架)
langgraph==1.0.4

# LLM 服务 & 数据抓取
litellm==1.30.0
ddgs==9.6.1
yfinance==0.2.66
finnhub-python==2.4.25
requests==2.32.5
beautifulsoup4==4.14.2
pandas==2.3.3
python-dotenv==1.1.1

# Tavily Search (AI 搜索 API)
tavily-python==0.3.0
langchain-tavily==0.2.13

# 性能优化
tenacity==9.1.2
pydantic==2.12.3
pydantic-settings==2.11.0

# 额外工具
openai==2.6.1
tiktoken==0.12.0
anthropic==0.71.0
docstring-parser==0.17.0
httpx-sse==0.4.3
dataclasses-json==0.6.7
orjson==3.11.4
ormsgpack==1.11.0
jsonpatch==1.33
jsonpointer==3.0.0
requests-toolbelt==1.0.0
xxhash==3.6.0
zstandard==0.25.0

# 数据库和缓存
SQLAlchemy==2.0.44
marshmallow==3.26.1
typing-inspect==0.9.0
mypy-extensions==1.1.0

# 监控和日志
colorama==0.4.6
tqdm==4.67.1
prometheus-client==0.21.0

# LangSmith 可观测性（可选）
langsmith>=0.1.0

# FastAPI & Web Server
fastapi==0.122.0
uvicorn[standard]==0.38.0
APScheduler==3.10.4

# Email Service
# 使用 Python 内置 smtplib，无需额外依赖

# PDF Export
reportlab==4.2.5
pypdf==4.2.0

# RAG - Vector Store & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Wikipedia (免费，无需 API key)
wikipedia==1.4.0
